+++ ims-apple/IMS Apple/Assets.xcassets/purpleGadfient.colorset/Contents.json
